Feedback Message Utility Functions

A set of helper functions for providing verbose feedback to the developer using this packages functions.

This function allows a package developer to automatically add these functions to their package.

use_feedback_helpers(pkg = ".")

msg_field(x)

msg_value(x)

msg_done(x)

msg_bullet(x, bullet = cli::symbol$bullet)

msg_err(x)

msg_path(x)

msg_info(x)

msg_code(x)

msg_feedback(x)

Arguments

pkg

Package - defaults to "."

x

The string passed to various msg_ functions.

bullet

What to use for the message's bullet. Defaults to cli::symbol$bullet

See also