htable Create a quick html table using my favorites options for DT::datatable()

htable Create a quick html table using my favorites options for DT::datatable()

quick_table(df, ...)

Arguments

df

data.frame to display

...

options sent through to DT::datatable

Examples

quick_table(mtcars)