Generalized Search and Search Engine Examples

Use these functions to search the web directly from your browser, building advanced queries and supplying common useful R related domains.

Query Github's internal search engine.

Query a search on rseek.org.

Query a search on the R Nabble Forum. Nabble is an innovative search engine for R messages.

search_online(s, query_url = "https://google.com/search?q=")

search_gh(
  s,
  type = "all",
  language = NULL,
  topic = NULL,
  user = NULL,
  org = NULL
)

search_rseek(s)

search_finzi(s)

search_nabble(s)

search_google(s)

search_rproject(s)

search_metacran(s)

Arguments

s
string to search for

query_url
string representing the URL to query; defaults to Google

type
what to search for, see details for options

language
optional language filter

topic
optional topic filter

user
optional user filter

org
optional org filter

Details

See https://cloud.r-project.org/search.html which showcases Google's advanced search feature to query only R-Project domain sites via the Google Search Engine.

Functions

search_gh:

search_rseek:

search_finzi:

search_nabble:

search_google:

search_rproject:

search_metacran:

References

http://www.sashagoodman.com/

http://finzi.psych.upenn.edu/search.html

http://finzi.psych.upenn.edu/search/manual.html#query

https://cloud.r-project.org/search.html

http://n4.nabble.com/help/Answer.jtp?id=31

Examples

# default search on google
search_online("polished.tech")

#> Error in search_online("polished.tech"): could not find function "search_online"

# search github (note: &ref=opensearch)
search_online("polished", "https://github.com/search?q=%s&ref=opensearch")

#> Error in search_online("polished", "https://github.com/search?q=%s&ref=opensearch"): could not find function "search_online"

# search Github with language:r, org:tychobra for 'polished' (note: '%3A' represents a ':')
search_online("polished", "https://github.com/search?q=%s+language%3Ar+org%3Atychobra")

#> Error in search_online("polished", "https://github.com/search?q=%s+language%3Ar+org%3Atychobra"): could not find function "search_online"

search_gh("websocket", language = "r", topic = "rshiny")

# search your org
search_gh("polished", org = "tychobra")